fix occasionally occurring error in fetch_waveform
The order and number of values returned by "WFMOutpre?" seem not to be consistent across different oscilloscopes (even not complying with the programmer's manual), so we explicitly give the order we want. On a Tektronix MSO71254C this caused an error at
trace.x_reference = int(float(pre[12]))
since pre[12] is 'V' instead of some numeric value.